+++ animation_nodes/data_structures/meshes/mesh_data.pyx
@@ -198,6 +198,7 @@ cdef class Mesh:
             newPolygons = triangulatePolygonsUsingEarClipMethod(self.vertices, polygons)
         self.edges = createValidEdgesList(polygons = newPolygons)
         self.polygons = newPolygons
+        self.derivedMeshDataCache.clear()
 
     def __repr__(self):
         return textwrap.dedent(