+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added *Calculate Loose Edges* option in *Mesh Object Output* node.
 - Added XoShiRo256 random number generators.
 - Added mesh *Triangulation* methods and *Triangulate Mesh* node.
+- Added Mesh Points Scatter node.
 
 ### Fixed